+++ src/aiida_aimall/controllers.py
@@ -23,7 +23,7 @@ class SmilesToGaussianController(FromGroupSubmissionController):
         group_label: the string of the group to put the GaussianCalculations in
         max_concurrent: maximum number of concurrent processes.
         code_label: label of code, e.g. gaussian@cedar
-        g16_opt_params: Dict of Gaussian parameters to use
+        gauss_opt_params: Dict of Gaussian parameters to use
         wfxgroup: group in which to store the resulting wfx files
         nprocs: number of processors for gaussian calculation
         mem_mb: amount of memory in MB for Gaussian calculation
@@ -45,7 +45,7 @@ class SmilesToGaussianController(FromGroupSubmissionController):
                 group_label = 'gaussianopt', # Resulting nodes will be in the gaussianopt group
                 max_concurrent = 1,
                 wfxgroup = "opt_wfx"
-                g16_opt_params = Dict(dict={
+                gauss_opt_params = Dict(dict={
                     'link0_parameters': {
                         '%chk':'aiida.chk',
                         "%mem": "4000MB",
@@ -69,18 +69,18 @@ class SmilesToGaussianController(FromGroupSubmissionController):
     group_label: str
     code_label: str
     max_concurrent: int
-    g16_opt_params: dict
+    gauss_opt_params: dict
     wfxgroup: str
     nprocs: int
     mem_mb: int
     time_s: int
 
-    WORKFLOW_ENTRY_POINT = "aimall.smitog16"
+    WORKFLOW_ENTRY_POINT = "aimall.smitogauss"
 
     def __init__(
         self,
         code_label: str,
-        g16_opt_params: dict,
+        gauss_opt_params: dict,
         wfxgroup: str,
         nprocs: int,
         mem_mb: int,
@@ -90,7 +90,7 @@ class SmilesToGaussianController(FromGroupSubmissionController):
     ):
         super().__init__(*args, **kwargs)
         self.code_label = code_label
-        self.g16_opt_params = g16_opt_params
+        self.gauss_opt_params = gauss_opt_params
         self.wfxgroup = wfxgroup
         self.nprocs = nprocs
         self.mem_mb = mem_mb
@@ -107,109 +107,13 @@ class SmilesToGaussianController(FromGroupSubmissionController):
         inputs = {
             "smiles": smiles,
             "gaussian_code": code,
-            "gaussian_parameters": Dict(self.g16_opt_params),
+            "gaussian_parameters": Dict(self.gauss_opt_params),
             "wfxgroup": Str(self.wfxgroup),
             "nprocs": Int(self.nprocs),
             "mem_mb": Int(self.mem_mb),
             "time_s": Int(self.time_s),
         }
         return inputs, WorkflowFactory(self.WORKFLOW_ENTRY_POINT)
-
-
-# class G16FragController(FromGroupSubmissionController):
-#     """A controller for submitting G16OptWorkChain
-
-#     Args:
-#         parent_group_label: the string of a group label which contains various structures as orm.Str nodes
-#         group_label: the string of the group to put the GaussianCalculations in
-#         max_concurrent: maximum number of concurrent processes.
-#         code_label: label of code, e.g. gaussian@cedar
-#         g16_opt_params: Dict of Gaussian parameters to use
-#         wfxgroup: group in which to store the resulting wfx files
-
-#     Returns:
-#         Controller object, periodically use run_in_batches to submit new results
-
-#     Note:
-#         In the typical use case, this is run on the outputs of the MultiFragmentWorkchain, which are by default added
-#             to the group inp_frag, so make sure `parent_group_label` matches that
-
-#     Example:
-#         In a typical use case of controllers, it is beneficial to check for new jobs periodically to submit.
-#             Either there may be new members of the parent_group to run, or some of the currently running jobs have run.
-#             So once a controller is defined, we can run it in a loop.
-
-#         ::
-
-#             controller = G16FragController(
-#                 code_label='gaussian@localhost',
-#                 parent_group_label = 'struct', # Add structures to run to struct group
-#                 group_label = 'gaussianopt', # Resulting nodes will be in the gaussianopt group
-#                 max_concurrent = 1,
-#                 wfxgroup = "opt_wfx"
-#                 g16_opt_params = Dict(dict={
-#                     'link0_parameters': {
-#                         '%chk':'aiida.chk',
-#                         "%mem": "4000MB",
-#                         "%nprocshared": 4,
-#                     },
-#                     'functional':'wb97xd',
-#                     'basis_set':'aug-cc-pvtz',
-#                     'charge': 0,
-#                     'multiplicity': 1,
-#                     'route_parameters': {'nosymmetry':None, 'Output':'WFX', 'opt':None, 'freq':None},
-#                     "input_parameters": {"output.wfx": None},
-#                     })
-#             )
-
-#             while True:
-#                 #submit Gaussian batches every hour
-#                 controller.submit_new_batch()
-#                 time.sleep(3600)
-
-#     """
-
-#     parent_group_label: str
-#     group_label: str
-#     code_label: str
-#     max_concurrent: int
-#     g16_opt_params: dict
-#     wfxgroup: str
-
-#     WORKFLOW_ENTRY_POINT = "aimall.g16opt"
-
-#     def __init__(
-#         self,
-#         code_label: str,
-#         g16_opt_params: dict,
-#         wfxgroup: str,
-#         *args,
-#         **kwargs,
-#     ):
-#         super().__init__(*args, **kwargs)
-#         self.code_label = code_label
-#         self.g16_opt_params = g16_opt_params
-#         self.wfxgroup = wfxgroup
-
-#     def get_extra_unique_keys(self):
-#         """Returns a tuple of extras keys in the order needed"""
-#         return ("smiles",)
-
-#     def get_inputs_and_processclass_from_extras(self, extras_values):
-#         """Constructs input for a GaussianWFXCalculation from extra_values
-
-#         Note: adjust the metadata options later for 6400MB and 7days runtime
-#         """
-#         code = orm.load_code(self.code_label)
-#         structure = self.get_parent_node_from_extras(extras_values)
-#         inputs = {
-#             "frag_label": Str(extras_values[0]),
-#             "fragment_dict": structure,
-#             "g16_code": code,
-#             "g16_opt_params": Dict(self.g16_opt_params),
-#             "wfxgroup": Str(self.wfxgroup),
-#         }
-#         return inputs, WorkflowFactory(self.WORKFLOW_ENTRY_POINT)
 
 
 class AimReorSubmissionController(FromGroupSubmissionController):
@@ -419,7 +323,7 @@ class GaussianSubmissionController(FromGroupSubmissionController):
         max_concurrent: maximum number of concurrent processes. Expected behaviour is to set to a large number
           since we will be submitting to Cedar which will manage
         code_label: label of code, e.g. gaussian@cedar
-        g16_sp_params: dictionary of parameters to use in gaussian calculation
+        gauss_sp_params: dictionary of parameters to use in gaussian calculation
 
     Returns:
         Controller object, periodically use run_in_batches to submit new results
@@ -444,7 +348,7 @@ class GaussianSubmissionController(FromGroupSubmissionController):
                 parent_group_label = 'struct', # Add structures to run to struct group
                 group_label = 'gaussiansp', # Resulting nodes will be in the gaussiansp group
                 max_concurrent = 1,
-                g16_sp_params = Dict(dict={
+                gauss_sp_params = Dict(dict={
                     'link0_parameters': {
                         '%chk':'aiida.chk',
                         "%mem": "4000MB",
@@ -470,7 +374,7 @@ class GaussianSubmissionController(FromGroupSubmissionController):
     group_label: str
     max_concurrent: int
     code_label: str
-    g16_sp_params: dict
+    gauss_sp_params: dict
     wfxgroup: str
     # GaussianWFXCalculation entry point as defined in aiida-aimall pyproject.toml
     CALCULATION_ENTRY_POINT = "gaussian"
@@ -478,7 +382,7 @@ class GaussianSubmissionController(FromGroupSubmissionController):
     def __init__(
         self,
         code_label: str,
-        g16_sp_params: dict,
+        gauss_sp_params: dict,
         wfxgroup: str,
         *args,
         **kwargs,
@@ -486,7 +390,7 @@ class GaussianSubmissionController(FromGroupSubmissionController):
         """Initialize the class, modifying with new values"""
         super().__init__(*args, **kwargs)
         self.code_label = code_label
-        self.g16_sp_params = g16_sp_params
+        self.gauss_sp_params = gauss_sp_params
         self.wfxgroup = wfxgroup
 
     # @validator("code_label")
@@ -515,7 +419,7 @@ class GaussianSubmissionController(FromGroupSubmissionController):
         structure = self.get_parent_node_from_extras(extras_values)
         inputs = {
             "code": code,
-            "parameters": Dict(self.g16_sp_params),
+            "parameters": Dict(self.gauss_sp_params),
             "structure": structure,
             "metadata": {
                 "options": {
