+++ src/aiida_aimall/data.py
@@ -2,7 +2,7 @@
 Data types provided by plugin
 
 Upon pip install, AimqbParameters  is accessible in AiiDA.data plugins
-Using the 'aimall' entry point
+Using the 'aimall.aimqb' entry point
 """
 
 from aiida.orm import Dict
