+++ tests/workchains/fixtures/aimall.smitogauss/default/aiida.chk
